+++ src/Rpc/RpcServer.cpp
@@ -513,7 +513,7 @@ bool RpcServer::f_on_block_json(const F_COMMAND_RPC_GET_BLOCK_DETAILS::request& 
 	uint64_t maxReward = 0;
 	uint64_t currentReward = 0;
 	int64_t emissionChange = 0;
-	size_t blockGrantedFullRewardZone = m_core.getCurrency().blockGrantedFullRewardZoneByBlockVersion(block_header.major_version);
+	uint64_t blockGrantedFullRewardZone = m_core.getCurrency().blockGrantedFullRewardZoneByBlockVersion(block_header.major_version);
 	res.block.effectiveSizeMedian = std::max(res.block.sizeMedian, blockGrantedFullRewardZone);
 
 	res.block.baseReward = blkDetails.baseReward;
