+++ src/Serialization/BlockchainExplorerDataSerialization.cpp
@@ -151,7 +151,7 @@ void serialize(TransactionDetails& transaction, ISerializer& serializer) {
 
   //serializer(transaction.signatures, "signatures");
   if (serializer.type() == ISerializer::OUTPUT) {
-    std::vector<std::pair<size_t, Crypto::Signature>> signaturesForSerialization;
+    std::vector<std::pair<uint64_t, Crypto::Signature>> signaturesForSerialization;
     signaturesForSerialization.reserve(transaction.signatures.size());
     size_t ctr = 0;
     for (const auto& signaturesV : transaction.signatures) {
@@ -160,15 +160,15 @@ void serialize(TransactionDetails& transaction, ISerializer& serializer) {
       }
       ++ctr;
     }
-    size_t size = transaction.signatures.size();
+    uint64_t size = transaction.signatures.size();
     serializer(size, "signaturesSize");
     serializer(signaturesForSerialization, "signatures");
   } else {
-    size_t size = 0;
+    uint64_t size = 0;
     serializer(size, "signaturesSize");
     transaction.signatures.resize(size);
 
-    std::vector<std::pair<size_t, Crypto::Signature>> signaturesForSerialization;
+    std::vector<std::pair<uint64_t, Crypto::Signature>> signaturesForSerialization;
     serializer(signaturesForSerialization, "signatures");
 
     for (const auto& signatureWithIndex : signaturesForSerialization) {
